Updated documentation, fixed some compilation warnings
work-items: #62916
git-tfs-id: [https://sede-pt-ce2.visualstudio.com/DefaultCollection]$/Basin Modeling/IBS/Trunk;C64416

+++ 3rdparty/NNlib/src/nncommon.cpp
@@ -135,7 +135,7 @@ void points_thingrid(int* pn, point** ppoints, int nx, int ny)
             double fi = (p->x - xmin) / stepx;
 
             if (fabs(rint(fi) - fi) < EPS)
-                i = rint(fi);
+                i = static_cast<int>( rint(fi) );
             else
                 i = (int) floor(fi);
         }
@@ -145,7 +145,7 @@ void points_thingrid(int* pn, point** ppoints, int nx, int ny)
             double fj = (p->y - ymin) / stepy;
 
             if (fabs(rint(fj) - fj) < EPS)
-                j = rint(fj);
+                j = static_cast<int>( rint(fj) );
             else
                 j = (int) floor(fj);
         }
